# openapi.yaml
openapi: 3.0.3
info:
  title: PixelProbe API
  # Version is dynamically set from version.py when served via /api/openapi.yaml
  version: 0.0.0
  description: |
    # PixelProbe REST API

    PixelProbe is a comprehensive media file corruption detection system that uses FFmpeg and other tools to scan and validate media files.

    Route and method coverage is checked in CI. Some response fields remain
    intentionally partial; consult docs/api.md and the response schemas before
    generating a strict client.

    ## Getting Started

    ### First-Time Setup

    1. Check if setup is required:
       ```bash
       curl -X GET http://your-server/api/auth/status
       ```

    2. If `first_run: true`, create admin account:
       ```bash
       curl -X POST http://your-server/api/auth/setup \
         -H "Content-Type: application/json" \
         -d '{"password": "your-secure-password"}'
       ```

    3. Login to get session cookie:
       ```bash
       curl -X POST http://your-server/api/auth/login \
         -H "Content-Type: application/json" \
         -d '{"username": "admin", "password": "your-secure-password"}' \
         -c cookies.txt
       ```

    ### Authentication Overview

    PixelProbe supports two authentication methods:

    #### Session-Based Authentication
    Used by the web interface. Login via `/auth/login` to receive a session cookie.

    #### API Token Authentication
    Recommended for programmatic access:

    1. Create an API token:
...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - sessionAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication using API tokens (opaque tokens generated by PixelProbe, not JWTs)
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication via login
  parameters:
    ProviderId:
      name: provider_id
      in: path
      required: true
      schema:
        type: integer
    RuleId:
      name: rule_id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    NotificationProviderInput:
      type: object
      required: [name, provider_type, configuration]
      properties:
        name:
          type: string
        provider_type:
          type: string
          enum: [email, ntfy, pushover, webhook]
        configuration:
          type: object
          description: Provider-specific credentials and destination settings.
          additionalProperties: true
        is_active:
          type: boolean
          default: true
    NotificationRuleInput:
      type: object
      required: [provider_id, event_type]
      properties:
        provider_id:
          type: integer
        event_type:
          type: string
          enum: [bitrot_suspected, scan_completed]
        conditions:
          type: object
          additionalProperties: true
          description: Event-specific condition values documented in docs/operational-evidence.md.
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
        is_active:
          type: boolean
          default: true
...
paths:
...
  /notifications/providers:
    get:
      summary: List notification providers
      tags: [Notifications]
      responses:
        '200': {description: Provider list}
    post:
      summary: Create a notification provider
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationProviderInput'
      responses:
        '201': {description: Provider created}
  /notifications/providers/{provider_id}:
    get:
      summary: Get a notification provider
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/ProviderId'
      responses:
        '200': {description: Provider details}
        '404': {description: Provider not found}
    put:
      summary: Update a notification provider
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/ProviderId'
      responses:
        '200': {description: Provider updated}
    delete:
      summary: Delete a notification provider
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/ProviderId'
      responses:
        '204': {description: Provider deleted}
  /notifications/providers/{provider_id}/test:
    post:
      summary: Send a notification-provider test
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/ProviderId'
      responses:
        '200': {description: Test dispatched}
  /notifications/rules:
    get:
      summary: List notification rules
      tags: [Notifications]
      responses:
        '200': {description: Rule list}
    post:
      summary: Create a notification rule
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRuleInput'
      responses:
        '201': {description: Rule created}
  /notifications/rules/{rule_id}:
    get:
      summary: Get a notification rule
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/RuleId'
      responses:
        '200': {description: Rule details}
        '404': {description: Rule not found}
    put:
      summary: Update a notification rule
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/RuleId'
      responses:
        '200': {description: Rule updated}
    delete:
      summary: Delete a notification rule
      tags: [Notifications]
      parameters:
        - $ref: '#/components/parameters/RuleId'
      responses:
        '204': {description: Rule deleted}
...